requirements.txt
django-admin==2.0.2
Django==5.0.0
djangorestframework==3.14.0
pytest==7.4.0
psycopg2-binary==2.9.6
numpy==1.21.5
Pillow==10.0.0
pytest-django==4.7.0
